+++ src/TableThermistor/TableThermistor.h
@@ -16,7 +16,7 @@ typedef struct
 class Thermistor
 {
 public:
-    Thermistor(byte thermistor_pin, const temperature_entry_t thermistor_table[], const uint8_t thermistor_table_length, float read_freq);
+    Thermistor(uint8_t thermistor_pin, const temperature_entry_t thermistor_table[], const uint8_t thermistor_table_length, float read_freq);
     ~Thermistor();
     void setReadFreq(float read_freq);
 
@@ -25,15 +25,15 @@ public:
 private:
     void ReadTemp();
 
-    byte pin;
+    uint8_t pin;
     const temperature_entry_t* tt;
     const uint8_t tt_len;
-    float T;
+    float   T;
     int16_t a;
 
-    float period; // in ms
-    boolean read_periodically;
-    unsigned long next_read_time; // in ms
+    float    period; // in ms
+    boolean  read_periodically;
+    uint32_t next_read_time; // in ms
 };
 
 #endif
